+++ scripts/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- `chain_id`: for Ethereum we now use `chain_id` instead of a string naming the network. 
+
 ## [0.8.2] - 2020-01-16
 
 ### Fixed